backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "assistant-stream>=0.0.30",
    "fastapi>=0.117.1",
    "pydantic-ai>=1.0.10",
    "python-dotenv>=1.1.1",
]
